1.管养列表填充服务器获取的数据，支持触摸横向滑动功能；2.修改养护记录登记页面，将附件及相关养护对象跳转按钮更改到标题栏右侧，上传按钮转移到底部，并可以跳转到相应活动。
