+++ app/src/main/java/com/nju/urbangreen/zhenjiangurbangreen/maintainRecord/MaintainListAdapter2.java
@@ -1,0 +1,8 @@
+package com.nju.urbangreen.zhenjiangurbangreen.maintainRecord;
+
+/**
+ * Created by tommy on 2017/8/17.
+ */
+
+public class MaintainListAdapter2 {
+}
